時間管理の誤差銃声

+++ Assets/ImitateDance/Scripts/Domain/Entity/Game/NoteLine/PositionEntity.cs
@@ -4,10 +4,10 @@ namespace ImitateDance.Scripts.Domain.Entity.Game.NoteLine
 {
     public sealed class PositionEntity
     {
-        public float NormalizePosition(float from, float to, float current)
+        public float NormalizePosition(float from, float to, double current)
         {
             var t = (current - from) / (to - from);
-            return Mathf.Clamp(t, 0, 1);
+            return Mathf.Clamp((float)t, 0, 1);
         }
     }
 }
